+++ src/mocks/index.ts
@@ -1,0 +1,7 @@
+export { companyFormSections } from '@/mocks/company'
+export { courierStats, couriers } from '@/mocks/couriers'
+export { senderStats, senders } from '@/mocks/senders'
+export {
+  shipmentStatuses,
+  shipmentStatusStats,
+} from '@/mocks/shipment-statuses'